fix: support gpt-5 reasoning in codememo eval
