+++ tests/evaluation/test_codememo_eval.py
@@ -41,6 +41,22 @@ class FakeSUT:
         self.reset_calls += 1
 
 
+class _FakeCompletions:
+    def __init__(self) -> None:
+        self.calls = []
+
+    def create(self, **kwargs):
+        self.calls.append(kwargs)
+        return SimpleNamespace(
+            choices=[SimpleNamespace(message=SimpleNamespace(content="ok"))]
+        )
+
+
+class _FakeClient:
+    def __init__(self) -> None:
+        self.chat = SimpleNamespace(completions=_FakeCompletions())
+
+
 def _write_project(root: Path, name: str = "project_demo") -> Path:
     project = root / name
     sessions = project / "sessions"
@@ -112,6 +128,47 @@ def test_run_evaluation_repeat_runs_records_per_run(tmp_path):
     assert {(row["run"], row["id"]) for row in detailed} == {
         (1, "q1"), (1, "q2"), (2, "q1"), (2, "q2")
     }
+
+
+def test_api_call_uses_legacy_chat_args_for_gpt4o():
+    client = _FakeClient()
+
+    result = codememo_eval._api_call_with_retry(
+        client,
+        [{"role": "user", "content": "hi"}],
+        max_tokens=77,
+        retries=1,
+        model="gpt-4o-mini",
+    )
+
+    assert result == "ok"
+    call = client.chat.completions.calls[0]
+    assert call["model"] == "gpt-4o-mini"
+    assert call["max_tokens"] == 77
+    assert call["temperature"] == 0.0
+    assert call["seed"] == 42
+    assert "max_completion_tokens" not in call
+
+
+def test_api_call_uses_gpt5_completion_arg_shape():
+    client = _FakeClient()
+
+    result = codememo_eval._api_call_with_retry(
+        client,
+        [{"role": "user", "content": "hi"}],
+        max_tokens=88,
+        retries=1,
+        model="gpt-5-mini",
+    )
+
+    assert result == "ok"
+    call = client.chat.completions.calls[0]
+    assert call["model"] == "gpt-5-mini"
+    assert call["max_completion_tokens"] == 88
+    assert call["reasoning_effort"] == "minimal"
+    assert "max_tokens" not in call
+    assert "temperature" not in call
+    assert "seed" not in call
 
 
 def test_run_evaluation_can_reset_working_memory_between_runs(tmp_path):
